WD-34911 - Fix dnd not working and empty revisions not appearing
